backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LangChain & LangGraph - let pip resolve compatible versions
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langgraph>=0.2.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# API Clients
httpx>=0.26.0
aiohttp>=3.9.1
amadeus>=12.0.0

# Task Queue (for async processing)
celery[redis]>=5.3.4
redis>=5.0.1

# Utilities
python-multipart>=0.0.6
email-validator>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
tenacity>=8.2.3

# Development
pytest>=7.4.4
pytest-asyncio>=0.23.3
black>=23.12.1
ruff>=0.1.11
mypy>=1.8.0
